+++ eslint.config.js
@@ -20,12 +20,6 @@ export default [
   ...unicorn,
   ...react,
   {
-    files: ['**/pdnd.cts'],
-    rules: {
-      'import/no-extraneous-dependencies': 'off',
-    },
-  },
-  {
     files: ['**/*.{ts,tsx,cts,mts}'],
     rules: {
       '@typescript-eslint/consistent-type-imports': 'error',
@@ -44,6 +38,8 @@ export default [
       '@typescript-eslint/no-misused-promises': 'off',
       '@typescript-eslint/restrict-plus-operands': 'off',
       'react/no-unstable-nested-components': 'off',
+      'react-hooks/refs': 'off',
+      'react-hooks/set-state-in-effect': 'off',
       'no-restricted-imports': [
         'error',
         {