+++ src/component/elements/ReactTable/Elements/ReactTableRow.tsx
@@ -82,7 +82,6 @@ function ReactTableRow(props: ReactTableRowProps) {
       highlight.hide();
     };
     // TODO: avoid this hack.
-    // eslint-disable-next-line react-hooks/react-compiler
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);