+++ src/component/hooks/useActiveSpectrumRangesViewState.ts
@@ -5,25 +5,33 @@ import { useChartData } from '../context/ChartContext.js';
 
 import { useActiveSpectrum } from './useActiveSpectrum.js';
 
-export const defaultRangesViewState: RangesViewState = {
-  showPeaks: false,
-  showMultiplicityTrees: false,
-  showIntegrals: false,
-  showIntegralsValues: true,
-  showJGraph: false,
-  displayingMode: 'spread',
-  integralsScaleRatio: 1,
-  showAssignmentsLabels: false,
-  showPublicationString: false,
-  showRanges: false,
-  publicationStringBounding: { x: 0, y: 0, width: 400, height: 0 },
-  rangesBounding: { x: 0, y: 0, width: 200, height: 150 },
-};
+export function getDefaultRangesViewState(nucleus: string): RangesViewState {
+  const isProton = nucleus === '1H';
+  const isCarbon = nucleus === '13C';
+
+  return {
+    showPeaks: isCarbon,
+    showMultiplicityTrees: false,
+    showIntegrals: false,
+    showIntegralsValues: isProton,
+    showJGraph: false,
+    displayingMode: 'spread',
+    integralsScaleRatio: 1,
+    showAssignmentsLabels: false,
+    showPublicationString: false,
+    showRanges: false,
+    publicationStringBounding: { x: 0, y: 0, width: 400, height: 0 },
+    rangesBounding: { x: 0, y: 0, width: 200, height: 150 },
+  };
+}
 
 export function useActiveSpectrumRangesViewState() {
   const activeSpectrum = useActiveSpectrum();
   const {
-    view: { ranges },
+    view: {
+      ranges,
+      spectra: { activeTab: nucleus },
+    },
   } = useChartData();
 
   const inset = useInsetOptions();
@@ -39,6 +47,6 @@ export function useActiveSpectrumRangesViewState() {
   ) {
     return ranges[activeSpectrum?.id];
   } else {
-    return defaultRangesViewState;
+    return getDefaultRangesViewState(nucleus);
   }
 }